+++ config/hook-convert/hook-convert.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+use Rector\Config\RectorConfig;
+
+return static function (RectorConfig $rectorConfig): void {
+    $rectorConfig->rule(\DrupalRector\Convert\Rector\HookConvertRector::class);
+};
